# Refactor rocm attention