Extract QKV projection and attention computation into pipelines. 

This will allow us to introduce ck flash attention in next PR