@ytaous 2023-02-27
What's the perf implication for this PR and also the PR with "flash attn" ?