@cloudhan 2023-02-28
@ytaous This refactoring will not affect the performace at all. Run a random model from benchmark suite shows it.

Before refactoring:
```
Average latency = 57.07 ms, Throughput = 1121.52 QPS

Latency(ms)	Latency_P50	Latency_P75	Latency_P90	Latency_P95	Latency_P99	Throughput(QPS)	model	graph_optimization_level	intra_op_num_threads	batch_size	sequence_length	test_cases	test_times	use_gpu
57.07	57.09	57.18	57.26	57.31	57.37	1121.52	fbv_v4_6L768_fp16_rocm.onnx	ENABLE_ALL	24	64	512	10	10	True
```

After refacoring:
```
Average latency = 56.95 ms, Throughput = 1123.72 QPS

Latency(ms)     Latency_P50     Latency_P75     Latency_P90     Latency_P95     Latency_P99     Throughput(QPS) model   graph_optimization_level        intra_op_num_threads    batch_size      sequence_length test_cases      test_times  use_gpu
56.95   56.99   57.08   57.16   57.19   57.21   1123.72 fbv_v4_6L768_fp16_rocm.onnx     ENABLE_ALL      24      64      512     10      10      True
```

NOTE:
1. runs on GPU0 GCD0.
2. with perf_determinism 1700