@cloudhan 2023-02-28
For the perf of the latter flash attention. Let's discuss it in that PR only.